Crate consensus_notifications

Structs

ConsensusCommitNotification

A commit notification to notify state sync of new commits.

ConsensusNotificationListener

The state sync component responsible for handling consensus requests and notifications.

ConsensusNotificationResponse

The result returned by state sync for a consensus notification.

ConsensusNotifier

The consensus component responsible for sending notifications and requests to state sync.

ConsensusSyncNotification

A commit notification to notify state sync to sync to the specified target.

Enums

ConsensusNotification

Error

Traits

ConsensusNotificationSender

The interface between state sync and consensus, allowing consensus to send synchronization notifications to state sync.

Functions

new_consensus_notifier_listener_pair

This method returns a (ConsensusNotifier, ConsensusNotificationListener) pair that can be used to allow consensus and state sync to communicate.